+++ EConnectApi/Helpers/DateTimeExtension.cs
@@ -4,7 +4,7 @@ namespace EConnectApi.Helpers
 {
     internal static class DateTimeExtension
     {
-        public static double ToJavaTimestamp(this DateTime dt)
+        public static long ToJavaTimestamp(this DateTime dt)
         {
             return DateTimeConverter.ToJavaTimestamp(dt);
         }
