+++ EConnectApi/Helpers/DoubleExtension.cs
@@ -4,7 +4,7 @@ namespace EConnectApi.Helpers
 {
     internal static class DoubleExtension
     {
-        public static DateTime ToDateTime(this double stamp)
+        public static DateTime ToDateTime(this long stamp)
         {
             return DateTimeConverter.ToDateTime(stamp);
         }
